Use more standard bindings in Hyprland
Although in my opinion the bindings before this commit are more
comfortable, they have a problem and that is that they are very unique
to my computer. I am usually using several computers (with Windows) and
I always confuse the bindings every time I change devices, so now, the
analog actions are more similar to the default bindings in Windows.