Remove UWSM
I am removing the use of this because it has been more of a drawback in
both usability and speed, although I wish to stick to the FreeDesktop
standard, I will have to stay away for the time being.

Perhaps in the future I will make a similar utility faster.